libkmod-signature: implement pkcs7 parsing with asn1c compiler
The patch adds pkcs7.asn1 description, taken from the linux kernel
and modified to be suitable for the asn1c compiler and a wrapper on
the asn1c's parser to make it more suitable for kmod.

Since kmod (modinfo) supports only one signature at the moment, the
wrapper ignores others and it also analyzes object IDs only for
hashing algorithms.

To compile the stuff you need to have asn1c installed, finish
./configure and invoke

make -C libkmod/pkcs7/asn1c-gen regen

to generate the parser.

The next patch will bring the generated files for the case of
absense of the asn1c compiler.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>